No error processing

+++ src/ClientBase.java
@@ -2,7 +2,7 @@ abstract class ClientBase {
   
   //    メンバ変数
   protected String clientName;  //  顧客名
-  protected int workCost;    //  仕事量
+  protected double workCost;    //  仕事量
   protected double workPeriod;  //  仕事期間
   protected String workType;    //  仕事内容
   
@@ -13,7 +13,7 @@ abstract class ClientBase {
   }
   
   abstract void setWorkCost();
-  abstract int getWorkCost();
+  abstract double getWorkCost();
   abstract void setWorkPeriod();
   abstract double getWorkPeriod();
   abstract void setWorkType();
